+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 - cta/CTA#770 - Reversed the order of asserts in testSubprocessWithStdinInput system test for more informative error messages
 - cta/CTA#781 - Fixed security issues with build pods
 - cta/CTA#772 - Make sonarcloud use yum.repos.d-public repos instead of the interal repos
+- cta/CTA#839 - Integer underflow when repacking tapes where number of files in storage class is less than number of existing tape copies
 
 ### Maintenance
 - cta/CTA#779 - Removed outdated doxygen files